fix: resolve all remaining issues — B2 nullable string, E3 Toml multiline, E5 FixExt, E6 WriteBytes

+++ PicoMsgPack/tests/PicoMsgPack.Tests/MsgPackReaderSequenceTests.cs
@@ -261,7 +261,7 @@ public class MsgPackReaderSequenceTests
             elen = extData.Length;
         }
         await Assert.That(tt).IsEqualTo(TokenType.Extension);
-        await Assert.That(tag).IsEqualTo(7);
+        await Assert.That((int)tag).IsEqualTo(7);
         await Assert.That(elen).IsEqualTo(1);
     }
 }
